refactor: 소켓 종료시 회의록 생성 #79
